+++ .gitignore
@@ -2,4 +2,4 @@
 .vscode/
 *__pycache__*
 config_db.py
-*test*
+fruitsfamily/test
