+++ fruitsfamily/static/carousel.css
