+++ fruitsfamily/static/carousel.js
